pyproject.toml
[project]
name = "finagent"
version = "0.1.3"
description = "Agentic trading and analysis system with pluggable agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "langchain-openai>=1.1.7",
    "langchain[openai]>=1.2.3",
    "langgraph>=1.0.6",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "prompt-toolkit>=3.0.0",
    "tinydb>=4.8.0",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
    "uvloop>=0.22.1",
    "yfinance>=1.0",
    "streamlit>=1.32.0",
    "reportlab>=4.0.0",
    "openpyxl>=3.1.0",
    "jinja2>=3.1.6",
    "weasyprint>=68.0",
    "web3>=7.14.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.1.0",
    "psycopg2-binary>=2.9.11",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
